Remove dependency on improper use of 'eval()' and use module conventions
