+++ consolidate.py
@@ -18,101 +18,105 @@ import sys
 import json
 from utilities import parse_config_arguments, make_identifier
 
-config_file, args = parse_config_arguments(sys.argv[1:])
+def consolidate(config, input_file, output_file):
+    print("consolidate: Starting consolidation")
 
-if config_file == None:
-    print("Usage: python consolidate.py --config=<config_file> <input_csv_file.csv>")
-    print("<input_csv_file.csv> is the output from take-inventory.py or extract-metadata.py and should be sorted by filename.")
-    print("<config_file> should be the same one given to take-inventory.py.")
-    sys.exit(2)
+    prefix = input_file.split('-')[0].lower()
+    terms = None
+    for content_set in config["inventory"]:
+        if content_set["name"].lower() == prefix:
+            terms = content_set["terms"]
+            break
 
-# Making the output filename assumes the input filename has only one .
-input_file = args[0]
-elements = input_file.split('.')
-output_file = elements[0] + '-consolidated.' + elements[1]
+    if terms is None:
+        print("Could not find terms for {}".format(prefix))
+        sys.exit(1)
 
-print("consolidate: Starting consolidation")
+    with open(input_file, encoding='utf-8') as f_in:
+        import csv    
+        reader = csv.reader(f_in)    
+        headers = next(reader)
+        
+        # Replace "Term" columns with individual terms; also remove Line and Extract (not meaningful)
+        index_term = headers.index("Term")
+        index_line = headers.index("Line")
+        index_extract = headers.index("Extract")    
+        headers.remove("Term")
+        headers.remove("Line")
+        headers.remove("Extract")
 
-config = None
-with open(config_file) as config_file:
-    config = json.load(config_file)
+        # Insert columns for each of the terms, plus a "Term_Total" column.
+        # All of these columns should be named with valid Python identifiers, which the make_identifier
+        # function guarantees.
 
-prefix = input_file.split('-')[0].lower()
-terms = None
-for content_set in config["inventory"]:
-    if content_set["name"].lower() == prefix:
-        terms = content_set["terms"]
-        break
+        for i in range(0, len(terms)):
+            headers.insert(index_term + i, make_identifier(terms[i]))
 
-if terms is None:
-    print("Could not find terms for {}".format(prefix))
-    sys.exit(1)
+        headers.insert(index_term + i + 1, "Term_Total")  # Note underscore in column name
 
-with open(input_file, encoding='utf-8') as f_in:
-    import csv    
-    reader = csv.reader(f_in)    
-    headers = next(reader)
-    
-    # Replace "Term" columns with individual terms; also remove Line and Extract (not meaningful)
-    index_term = headers.index("Term")
-    index_line = headers.index("Line")
-    index_extract = headers.index("Extract")    
-    headers.remove("Term")
-    headers.remove("Line")
-    headers.remove("Extract")
+        with open(output_file, 'w', encoding='utf-8', newline='') as f_out:        
+            writer = csv.writer(f_out)
+            writer.writerow(headers)
 
-    # Insert columns for each of the terms, plus a "Term_Total" column.
-    # All of these columns should be named with valid Python identifiers, which the make_identifier
-    # function guarantees.
-
-    for i in range(0, len(terms)):
-        headers.insert(index_term + i, make_identifier(terms[i]))
-
-    headers.insert(index_term + i + 1, "Term_Total")  # Note underscore in column name
-
-    with open(output_file, 'w', encoding='utf-8', newline='') as f_out:        
-        writer = csv.writer(f_out)
-        writer.writerow(headers)
-
-        term_counts = [0] * len(terms)
-        current_row = next(reader)
-
-        while current_row is not None:
-            if current_row == None:
-                break
-
-            next_row = next(reader, None)  # Assigns None if we're at the end
-
-            term = current_row[index_term]
-            term_counts[terms.index(term)] += 1
-
-            # If the filename in the next row is the same, continue with the next line
-            if next_row != None and current_row[1] == next_row[1]:
-                current_row = next_row
-                continue;
-
-            # Otherwise, write the term count columns
-            # First, remove the Extract, Line, and Term columns, which are no longer relevant.
-            # We do this in reverse order because we're using indices. Note that this does assume
-            # the ordering generated by take-inventory.py and extract-metadata.py.
-            current_row.pop(index_extract)
-            current_row.pop(index_line)
-            current_row.pop(index_term)
-
-            total = 0
-
-            for i in range(0, len(terms)):
-                current_row.insert(index_term + i, term_counts[i])                
-                total += term_counts[i]
-
-            # Add one more row with the total count of terms, which accomodates sorting
-            current_row.insert(index_term + i + 1, total)
-
-            # Write the row
-            writer.writerow(current_row)
-
-            # Reset state for the next round, including the counts
-            current_row = next_row
             term_counts = [0] * len(terms)
+            current_row = next(reader)
 
-print("consolidate: Consolidation complete")
+            while current_row is not None:
+                if current_row == None:
+                    break
+
+                next_row = next(reader, None)  # Assigns None if we're at the end
+
+                term = current_row[index_term]
+                term_counts[terms.index(term)] += 1
+
+                # If the filename in the next row is the same, continue with the next line
+                if next_row != None and current_row[1] == next_row[1]:
+                    current_row = next_row
+                    continue
+
+                # Otherwise, write the term count columns
+                # First, remove the Extract, Line, and Term columns, which are no longer relevant.
+                # We do this in reverse order because we're using indices. Note that this does assume
+                # the ordering generated by take-inventory.py and extract-metadata.py.
+                current_row.pop(index_extract)
+                current_row.pop(index_line)
+                current_row.pop(index_term)
+
+                total = 0
+
+                for i in range(0, len(terms)):
+                    current_row.insert(index_term + i, term_counts[i])                
+                    total += term_counts[i]
+
+                # Add one more row with the total count of terms, which accomodates sorting
+                current_row.insert(index_term + i + 1, total)
+
+                # Write the row
+                writer.writerow(current_row)
+
+                # Reset state for the next round, including the counts
+                current_row = next_row
+                term_counts = [0] * len(terms)
+
+    print("consolidate: Consolidation complete")
+
+if __name__ == "__main__":
+    config_file, args = parse_config_arguments(sys.argv[1:])
+
+    if config_file == None:
+        print("Usage: python consolidate.py --config=<config_file> <input_csv_file.csv>")
+        print("<input_csv_file.csv> is the output from take-inventory.py or extract-metadata.py and should be sorted by filename.")
+        print("<config_file> should be the same one given to take-inventory.py.")
+        sys.exit(2)
+
+    # Making the output filename assumes the input filename has only one .
+    input_file = args[0]
+    elements = input_file.split('.')
+    output_file = elements[0] + '-consolidated.' + elements[1]
+
+    config = None
+    with open(config_file) as config_file:
+        config = json.load(config_file)
+
+    consolidate(config, input_file, output_file)
